# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "CatalogSearchRowResponse": {
        "additionalProperties": false,
        "properties": {
          "collector_number": {
            "title": "Collector Number",
            "type": "string"
          },
          "finishes": {
            "description": "Canonical finish values: normal, foil, etched.",
            "items": {
              "enum": [
                "normal",
                "foil",
                "etched"
              ],
              "type": "string"
            },
            "title": "Finishes",
            "type": "array"
          },
          "image_uri_normal": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Uri Normal"
          },
          "image_uri_small": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Image Uri Small"
          },
          "lang": {
            "description": "Catalog language code. Common values include: en, ja, de, fr, it, es, pt, ru, ko, zhs, zht, ph.",
            "title": "Lang",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "rarity": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rarity"
          },
          "scryfall_id": {
            "title": "Scryfall Id",
            "type": "string"
          },
          "set_code": {
            "title": "Set Code",
            "type": "string"
          },
          "set_name": {
            "title": "Set Name",
            "type": "string"
          },
          "tcgplayer_product_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tcgplayer Product Id"
          }
        },
        "required": [
          "scryfall_id",
          "name",
          "set_code",
          "set_name",
          "collector_number",
          "lang",
          "rarity",
          "finishes",
          "tcgplayer_product_id",
          "image_uri_small",
          "image_uri_normal"
        ],
        "title": "CatalogSearchRowResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "MTG Inventory Tool API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/cards/oracle/{oracle_id}/printings": {
      "get": {
        "operationId": "card_printings_lookup_cards_oracle__oracle_id__printings_get",
        "parameters": [
          {
            "in": "path",
            "name": "oracle_id",
            "required": true,
            "schema": {
              "title": "Oracle Id",
              "type": "string"
            }
          },
          {
            "description": "Catalog search language filter. Recommended codes include: en, ja, de, fr, it, es, pt, ru, ko, zhs, zht, ph. Search currently accepts the raw stored catalog language codes. Omit this parameter to prefer English printings by default. Use `all` to include every available catalog language.",
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Catalog search language filter. Recommended codes include: en, ja, de, fr, it, es, pt, ru, ko, zhs, zht, ph. Search currently accepts the raw stored catalog language codes. Omit this parameter to prefer English printings by default. Use `all` to include every available catalog language.",
              "title": "Lang"
            }
          },
          {
            "description": "Catalog scope to search. Omit this parameter or use `default` for the mainline card-add flow. Use `all` to include auxiliary catalog objects such as tokens, emblems, and art-series rows.",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Catalog scope to search. Omit this parameter or use `default` for the mainline card-add flow. Use `all` to include auxiliary catalog objects such as tokens, emblems, and art-series rows.",
              "enum": [
                "default",
                "all"
              ],
              "title": "Scope"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogSearchRowResponse"
                  },
                  "title": "Response Card Printings Lookup Cards Oracle  Oracle Id  Printings Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Schema not ready"
          }
        },
        "summary": "Card Printings Lookup"
      }
    },
    "/cards/search": {
      "get": {
        "operationId": "cards_search_cards_search_get",
        "parameters": [
          {
            "in": "query",
            "name": "query",
            "required": true,
            "schema": {
              "title": "Query",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "set_code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Set Code"
            }
          },
          {
            "in": "query",
            "name": "rarity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Rarity"
            }
          },
          {
            "description": "Accepted input values: normal, nonfoil, foil, etched. Canonical response values: normal, foil, etched. The input alias `nonfoil` is normalized to `normal`.",
            "in": "query",
            "name": "finish",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "normal",
                    "nonfoil",
                    "foil",
                    "etched"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Accepted input values: normal, nonfoil, foil, etched. Canonical response values: normal, foil, etched. The input alias `nonfoil` is normalized to `normal`.",
              "title": "Finish"
            }
          },
          {
            "description": "Catalog search language filter. Recommended codes include: en, ja, de, fr, it, es, pt, ru, ko, zhs, zht, ph. Search currently accepts the raw stored catalog language codes.",
            "in": "query",
            "name": "lang",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Catalog search language filter. Recommended codes include: en, ja, de, fr, it, es, pt, ru, ko, zhs, zht, ph. Search currently accepts the raw stored catalog language codes.",
              "title": "Lang"
            }
          },
          {
            "description": "Catalog scope to search. Omit this parameter or use `default` for the mainline card-add flow. Use `all` to include auxiliary catalog objects such as tokens, emblems, and art-series rows.",
            "in": "query",
            "name": "scope",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Catalog scope to search. Omit this parameter or use `default` for the mainline card-add flow. Use `all` to include auxiliary catalog objects such as tokens, emblems, and art-series rows.",
              "enum": [
                "default",
                "all"
              ],
              "title": "Scope"
            }
          },
          {
            "in": "query",
            "name": "exact",
            "required": false,
            "schema": {
              "default": false,
              "title": "Exact",
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CatalogSearchRowResponse"
                  },
                  "title": "Response Cards Search Cards Search Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Schema not ready"
          }
        },
        "summary": "Cards Search"
      }
    },
...
    "/inventories/{inventory_slug}/audit": {
      "get": {
        "operationId": "inventory_audit_list_inventories__inventory_slug__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "inventory_slug",
            "required": true,
            "schema": {
              "title": "Inventory Slug",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "item_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Item Id"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/InventoryAuditEventResponse"
                  },
                  "title": "Response Inventory Audit List Inventories  Inventory Slug  Audit Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Validation error"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Authentication required"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Not found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Internal server error"
          },
          "503": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorResponse"
                }
              }
            },
            "description": "Schema not ready"
          }
        },
        "summary": "Inventory Audit List"
      }
    },
...
  }
}